:construction: (jaewoo/feature/back-end/#53) PostDetailData 파일을 Code Convention에 맞게 Refactoring
